+++ src/components/search/search.jsx
@@ -6,16 +6,16 @@ import omit from '../../utilities/omit';
 import Flag from '../flag';
 import Spinner from '../spinner';
 import SearchIcon from '../icon/icons/search';
-// import './search.scss';
+import SearchStyles from './search-styles';
 
-function renderDevelopment(development) {
+function renderDevelopment(development, classes) {
   if (!development) {
     return null;
   }
 
-  const developmentClass = classNames('search__result-development', {
-    'search__result-development--positive': development > 0,
-    'search__result-development--negative': development < 0,
+  const developmentClass = classNames(classes.development, {
+    'development--positive': development > 0,
+    'development--negative': development < 0,
   });
 
   const positiveArrowPath = 'M0,16 L16,16 L8,0';
@@ -37,8 +37,8 @@ function renderDevelopment(development) {
 }
 
 class Search extends React.Component {
-  constructor(props) {
-    super(props);
+  constructor(props, context) {
+    super(props, context);
 
     this.state = {
       value: props.value,
@@ -51,6 +51,7 @@ class Search extends React.Component {
     this.handleClick = this.handleClick.bind(this);
     this.handleClickOutside = this.handleClickOutside.bind(this);
     this.renderResult = this.renderResult.bind(this);
+    this.classes = context.styleManager.render(SearchStyles);
   }
 
   componentDidMount() {
@@ -117,16 +118,16 @@ class Search extends React.Component {
     const Wrapper = href ? 'a' : 'span';
 
     return (
-      <li className="search__result" key={kebabCase(name + market)}>
+      <li className={this.classes.result} key={kebabCase(name + market)}>
         <Wrapper href={href}>
-          <div className="search__result-info">
-            <div className="search__result-name">{ name }</div>
-            <div className="search__result-market">
+          <div className={this.classes.info}>
+            <div className={this.classes.name}>{ name }</div>
+            <div className={this.classes.market}>
               <Flag countryCode={countryCode} />
               <span>{ market }</span>
             </div>
           </div>
-          { renderDevelopment(development) }
+          { renderDevelopment(development, this.classes) }
         </Wrapper>
       </li>
     );
@@ -140,18 +141,18 @@ class Search extends React.Component {
     }
 
     const searchResultsClass = classNames('search__results', {
-      'search__results--left': alignResults === 'left',
-      'search__results--right': alignResults === 'right',
+      left: alignResults === 'left',
+      right: alignResults === 'right',
     });
 
     const noResults = (
-      <li className="search__result search__result--no-results">
+      <li className={classNames(this.classes.result, this.classes.noResults)}>
         { this.props.noResults }
       </li>
     );
 
     const spinner = (
-      <li className="search__result search__result--is-loading">
+      <li className={classNames(this.classes.result, this.classes.isLoading)}>
         <Spinner />
       </li>
     );
@@ -175,7 +176,7 @@ class Search extends React.Component {
   render() {
     const { className, placeholder, isLoading, results, ...rest } = this.props;
     const { value, showResults } = this.state;
-    const classes = classNames('search', className);
+    const classes = classNames(this.classes.search, className);
     const show = (results || isLoading) && showResults && value.length > 0;
 
     return (
@@ -183,7 +184,7 @@ class Search extends React.Component {
         <input
           {...omit(rest, 'search', 'results', 'noResults', 'searchDebounceWait', 'isLoading', 'alignResults',
             'disableOnClickOutside', 'enableOnClickOutside', 'showResults', 'showNoResults', 'resultRenderer')}
-          className="search__input"
+          className={this.classes.input}
           type="search"
           placeholder={placeholder}
           onChange={this.onChange}
@@ -233,6 +234,10 @@ Search.propTypes = {
   isLoading: PropTypes.bool,
   value: PropTypes.string,
   alignResults: PropTypes.oneOf(['left', 'right']),
+};
+
+Search.contextTypes = {
+  styleManager: PropTypes.object.isRequired,
 };
 
 export default Search;