+++ src/components/table/table.jsx
@@ -1,19 +1,13 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
-// TODO: Move SCSS into JSS
-// import './table.scss';
+import TableStyles from './table-styles';
 
-function Table(props) {
-  const { className, children, size, ...rest } = props;
-  const classes = classNames('table', {
-    'table--xs': size === 'xs',
-    'table--sm': size === 'sm',
-    'table--md': size === 'md',
-    'table--lg': size === 'lg',
-  }, className);
+function Table({ className, children, size, ...rest }, { styleManager }) {
+  const classes = styleManager.render(TableStyles);
+  const usedClassName = classNames(classes.table, size, className);
 
   return (
-    <table {...rest} className={classes}>
+    <table {...rest} className={usedClassName}>
       { children }
     </table>
   );
@@ -27,6 +21,10 @@ Table.propTypes = {
   className: PropTypes.string,
   children: PropTypes.node,
   size: PropTypes.oneOf(['xs', 'sm', 'md', 'lg']),
+};
+
+Table.contextTypes = {
+  styleManager: PropTypes.object.isRequired,
 };
 
 export default Table;