+++ src/components/tfoot/tfoot-styles.jsx
@@ -1,0 +1,9 @@
+import { createStyleSheet } from 'jss-theme-reactor';
+import styleUtils from '../table/style-utilities';
+
+export default createStyleSheet('Tfoot', () => ({
+  tfoot: {
+    display: 'block',
+    ...styleUtils.sizes(),
+  },
+}));