+++ src/components/thead/thead.jsx
@@ -1,7 +1,6 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
-// TODO: Move SCSS into JSS
-// import './thead.scss';
+import TheadStyles from './thead-styles';
 
 function Thead({
   className,
@@ -9,18 +8,12 @@ function Thead({
   size,
   variant,
   ...rest
-}) {
-  const classes = classNames('thead', {
-    'thead--xs': size === 'xs',
-    'thead--sm': size === 'sm',
-    'thead--md': size === 'md',
-    'thead--lg': size === 'lg',
-    'thead--primary': variant === 'primary',
-    'thead--secondary': variant === 'secondary',
-  }, className);
+}, { styleManager }) {
+  const classes = styleManager.render(TheadStyles);
+  const usedClassName = classNames(classes.thead, size, (variant ? [variant] : []), className);
 
   return (
-    <thead {...rest} className={classes}>
+    <thead {...rest} className={usedClassName}>
       { children }
     </thead>
   );
@@ -33,6 +26,10 @@ Thead.propTypes = {
   children: PropTypes.node,
   size: PropTypes.oneOf(['xs', 'sm', 'md', 'lg']),
   variant: PropTypes.oneOf(['primary', 'secondary']),
+};
+
+Thead.contextTypes = {
+  styleManager: PropTypes.object.isRequired,
 };
 
 export default Thead;