+++ src/components/widget/widget.jsx
@@ -1,31 +1,32 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
-// import './widget.scss';
+import WidgetStyles from './widget-styles';
 
-function renderHeader(header) {
+function renderHeader(header, classes) {
   let headerBody = header;
 
   if (typeof header === 'string') {
-    headerBody = <h4 className="widget__heading">{ header }</h4>;
+    headerBody = <h4 className={classes.heading}>{ header }</h4>;
   }
 
   return (
-    <div className="widget__header">
+    <div>
       { headerBody }
     </div>
   );
 }
 
-export default function Widget({ header, children, fullWidth, className, ...rest }) {
-  const classes = classNames('widget', className);
-  const bodyClasses = classNames('widget__body', {
-    'widget__body--full-width': fullWidth,
-    'widget__body--no-header': !header,
+export default function Widget({ header, children, fullWidth, className, ...rest }, { styleManager }) {
+  const classes = styleManager.render(WidgetStyles);
+  const usedClassName = classNames(classes.widget, className);
+  const bodyClasses = classNames(classes.body, {
+    [classes.fullWidth]: fullWidth,
+    [classes.noHeader]: !header,
   });
 
   return (
-    <div {...rest} className={classes}>
-      { header ? renderHeader(header) : null }
+    <div {...rest} className={usedClassName}>
+      { header ? renderHeader(header, classes) : null }
       <div className={bodyClasses}>
         { children }
       </div>
@@ -42,4 +43,8 @@ Widget.propTypes = {
   header: PropTypes.node,
   className: PropTypes.string,
   fullWidth: PropTypes.bool,
+};
+
+Widget.contextTypes = {
+  styleManager: PropTypes.object.isRequired,
 };