+++ styleguide.config.js
@@ -7,15 +7,41 @@ const config = require('./webpack.config')('styleguidist');
 const dir = path.join(__dirname, 'src');
 
 const portedComponents = [
-  'badge', 'labeled-value', 'icon', 'logo', 'flag', 'avatar',
-  'spark-graph', 'radio-group', 'input', 'li', 'ul',
+  'badge',
+  'labeled-value',
+  'icon',
+  'logo',
+  'flag',
+  'avatar',
+  'radio-group',
+  'input',
+  'legend',
+  'dropdown',
+  'pane',
+  'horizontal-nav',
+  'tooltip',
+  'button',
+  'search',
+  'table',
+  'tbody',
+  'thead',
+  'tfoot',
+  'tr',
+  'th',
+  'td',
+  'widget',
+  'spinner',
+  'ratio-bar',
+  'alert',
+  'li',
+  'ul',
 ];
 
 const quickFixedComponents = [
-  'table', 'tbody', 'thead', 'tfoot', 'tr', 'th', 'td',
-  'button', 'graph-tooltip', 'horizontal-nav', 'legend', 'pane',
-  'dropdown', 'alert', 'nav-bar', 'tooltip', 'search',
-  'spinner', 'widget', 'range-selector', 'ratio-bar', 'select',
+  'graph-tooltip',
+  'nav-bar',
+  'select',
+  'range-selector',
 ];
 
 const allComponents = [...portedComponents, ...quickFixedComponents];